requirements.txt
Django
mysqlclient>=1.3.13
django-cors-headers
djangorestframework
django-rest-auth
django-allauth
faker